refactor: introduce info file schema v2.5.0 for the assertions field
Adding the assertions field to the existing v2.4.0 schema would have
changed a released schema in place. Follow the established convention
instead: new version file, migrator from v2.4.0, and updated example
collections.
