+++ examples/collection/collection.json
@@ -1,7 +1,7 @@
 {
   "id": "dbba497e-b5a1-47f3-86d4-5461c01e034f",
   "title": "New Collection",
-  "version": "2.4.0",
+  "version": "2.5.0",
   "variables": {
     "test": {
       "value": "I contain text",
