+++ examples/collection/echo/request.json
@@ -1,7 +1,7 @@
 {
   "id": "35efb7a3-84ab-4d28-a91c-2945c0591c6e",
   "title": "Echo",
-  "version": "2.4.0",
+  "version": "2.5.0",
   "url": {
     "base": "https://echo.free.beeceptor.com",
     "query": []
