+++ src/main/persistence/service/info-files/latest.ts
@@ -4,11 +4,11 @@ import { Collection } from 'shim/objects/collection';
 import { Folder } from 'shim/objects/folder';
 import { TrufosRequest } from 'shim/objects/request';
 import { VariableMap } from 'shim/objects/variables';
-import { InfoFile, VERSION, RequestInfoFile, FolderInfoFile, CollectionInfoFile } from './v2-4-0';
+import { InfoFile, VERSION, RequestInfoFile, FolderInfoFile, CollectionInfoFile } from './v2-5-0';
 import { SettingsService } from '../settings-service';
 
 export { createGitIgnore, GIT_IGNORE_FILE_NAME } from './v2-0-0';
-export { VERSION, InfoFile, CollectionInfoFile, FolderInfoFile, RequestInfoFile } from './v2-4-0';
+export { VERSION, InfoFile, CollectionInfoFile, FolderInfoFile, RequestInfoFile } from './v2-5-0';
 
 /**
  * Deep clones the given object and removes any properties that are not allowed in an info file.
