+++ src/main/persistence/service/info-files/migrators.ts
@@ -11,6 +11,7 @@ import { InfoFileMigrator as V2_1_0 } from './v2-1-0';
 import { InfoFileMigrator as V2_2_0 } from './v2-2-0';
 import { InfoFileMigrator as V2_3_0 } from './v2-3-0';
 import { InfoFileMigrator as V2_4_0 } from './v2-4-0';
+import { InfoFileMigrator as V2_5_0 } from './v2-5-0';
 import { SemVer } from 'main/util/semver';
 
 // add new migrators here
@@ -25,6 +26,7 @@ const MIGRATOR_ARRAY = [
   new V2_2_0(),
   new V2_3_0(),
   new V2_4_0(),
+  new V2_5_0(),
 ];
 
 const MIGRATORS = new Map<string, AbstractInfoFileMigrator<VersionedObject, VersionedObject>>(
